# BE/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jasindo User Management API",
        "description": "CRUD API for the Jasindo front end coding test. Auth via a JWT carried in an HttpOnly, SameSite=Lax cookie -- Swagger's \"Try it out\" cannot authenticate directly because the cookie is invisible to JavaScript by design; log in from the app itself, then Swagger inherits the cookie on same-origin requests.",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001/api/v1",
            "description": "Local development"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Exchange credentials for a session cookie",
                "operationId": "67d6d0fdd54fba132e1b6f55ca254183",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@jasindo.test"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authenticated. Sets the jasindo_token HttpOnly cookie."
                    },
                    "422": {
                        "description": "Invalid credentials or missing fields"
                    },
                    "429": {
                        "description": "Too many login attempts"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Clear the session cookie",
                "operationId": "28d48c13fca984400ac8de99f54ee13a",
                "responses": {
                    "200": {
                        "description": "Cookie cleared"
                    },
                    "401": {
                        "description": "Not authenticated"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Current authenticated user, and silently renews the session",
                "operationId": "bc2032a379aa97707ccc61511897c120",
                "responses": {
                    "200": {
                        "description": "The authenticated user"
                    },
                    "401": {
                        "description": "Not authenticated"
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "description": "HttpOnly JWT cookie set by POST /login. Never readable from JavaScript.",
                "name": "jasindo_token",
                "in": "cookie"
            }
        }
    },
...
}